wip(discord): Document interface properties & move to types.ts
